Adding separate ovpncli lib
