+++ test/ovpncli/cli.cpp
@@ -1634,30 +1634,3 @@ int openvpn_client(int argc, char *argv[], const std::string *profile_content)
     return ret;
 }
 
-#ifndef OPENVPN_OVPNCLI_OMIT_MAIN
-
-int main(int argc, char *argv[])
-{
-    int ret = 0;
-
-#ifdef OPENVPN_LOG_LOGBASE_H
-    LogBaseSimple log;
-#endif
-
-#if defined(OPENVPN_PLATFORM_WIN)
-    SetConsoleOutputCP(CP_UTF8);
-#endif
-
-    try
-    {
-        ret = openvpn_client(argc, argv, nullptr);
-    }
-    catch (const std::exception &e)
-    {
-        std::cout << "Main thread exception: " << e.what() << std::endl;
-        ret = 1;
-    }
-    return ret;
-}
-
-#endif
