+++ test/ovpncli/cli.hpp
@@ -1,0 +1,4 @@
+#pragma once
+#include <string>
+
+int openvpn_client(int argc, char *argv[], const std::string *profile_content);
